+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.3.0 [2018-05-14]
+### Notes
+- Since I am leaving the Desktop Support group, this will likely be the last major release. I will try to update SNAFU as I can when things bring or change, but SNAFU will no longer be in active development.
+
+### Added
+- Equipment Staging Transfer label manual print.
+
 ## 1.2.26 [2018-04-13]
 ### Fixed
 - Small bug with incorrect ticket status being sent via keyboard shortcuts.